Move EDGE setup into tools
